refactor: rename BorgClient to App
